run_make_support: rename `cygpath_windows` to `get_windows_path` and move under `external_deps` as private
