+++ src/tools/run-make-support/src/external_deps/mod.rs
@@ -1,10 +1,5 @@
 //! This module contains external tool dependencies that we assume are available in the environment,
 //! such as `cc` or `python`.
-//!
-//! # Notes
-//!
-//! - This is not the *only* place where external dependencies are assumed or referenced. For
-//!   example, see [`cygpath_windows`][crate::path_helpers::cygpath_windows].
 
 pub mod c_build;
 pub mod cc;
@@ -14,3 +9,6 @@ pub mod llvm;
 pub mod python;
 pub mod rustc;
 pub mod rustdoc;
+
+// Library-internal external dependency.
+mod cygpath;
